test(e2e): update send and secure send tests (#4804)
### Description
This should speed up CI runs and reduce flake. Times were taken from the
following action runs:
-
https://github.com/valora-inc/wallet/actions/runs/7643492040/job/20831482604
(Before)
-
https://github.com/valora-inc/wallet/actions/runs/7660599651/job/20878225247
(After)

| iOS Duration Before | iOS Duration After | Android Duration Before |
Android Duration After |
| ---- | ---- | ---- | ---- |
| 45m | 32m | 36m | 25m | 

- Splits up the existing send spec into 3 different specs:
`Send.spec.js`, `SendOld.spec.js`, and `SecureSend.spec.js`; this has
the advantage of farming out specs to available devices and prevents one
failing send test from from causing a rerun of all the send tests.
- Uses a random comment during the send flow and validates that the
transaction was successful by checking for the comment in the
transaction feed. Previously tests only asserted that the user landed in
the transaction feed.
- Sets the `When multi-token send flow to recent recipient (new flow)`
test to use a new instance. A new instance with launch is similar to
reloadReactNative where the app state is as if the user completely
closed and reopened the app instead of just backgrounded.

### Test plan

- Tested locally on iOS
- Tested in CI on Android and iOS

### Related issues

N/A

### Backwards compatibility

Yes