chore(vulnerability): Ignore dependency vulnerability GHSA-78xj-cgh5-2h22 (#4888)
### Description

As of right now there is no fix and also this is a server-side
vulnerability so adding this to the ignore list

[Github advisory
](https://github.com/advisories/GHSA-78xj-cgh5-2h22)

> An issue in NPM IP Package v.1.1.8 and before allows an attacker to
execute arbitrary code and obtain sensitive information via the
isPublic() function. This can lead to potential Server-Side Request
Forgery (SSRF) attacks. The core issue is the function's failure to
accurately distinguish between public and private IP addresses.


### Test plan

n/a

### Related issues

n/a

### Backwards compatibility

n/a