ci(e2e): test changes for xcode 15 support (#4955)
### Description

Adds some custom logic to set the default currency and language on e2e
test runners. This allows consistency regardless of the device e2e tests
are run on e.g. personal dev devices, remote ci machines, etc. While the
desirable way to achieve this would be setting the simulator location as
in 60a90d7bae690f94e131c3a4907b30b2f141ba4a or with Detox's
[`device.setLocation(...)`](https://wix.github.io/Detox/docs/next/api/device/#devicesetlocationlat-lon)
it appears that when setting a simulator location in Xcode 15.2 this is
not respected and the machine's physical location is used instead.

Alternatively we could do something like this
https://stackoverflow.com/a/75914924, but that seems more opaque for
when e2e tests behavior deviates.

### Test plan

- Tested in CI on existing Android Runners & new Mac Runner.

### Related issues

- Related ACT-1099

### Backwards compatibility

- Yes - only impacts e2e test runs.

### Network scalability

- N/A