feat: add blocking max amount warning for jumpstart send flow (#5009)
### Description

As the title. This warning could be shown in combination with any other
warnings about the prepared transactions that are there in the
EnterAmount component, which I think is acceptable.

### Test plan

With the max value set at $1USD:


https://github.com/valora-inc/wallet/assets/20150449/83634079-1f12-4e7b-8fc4-3770ecf13da7



### Related issues

- Fixes RET-993

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)

---------

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>