feat: support fee currency adapters (#5010)
### Description

This adds support for fee currency adapters.

Everything seems to be working fine.
I made sure to update all flows using viem (send, swap, WC) to work with
adapters (correct token and decimals used mostly).

One problem though is sent transactions with a fee currency adapter set
aren't returned by blockchain-api. I'll investigate this. In the
screenshots below, the sent TX is visible but is only due to the local
state and won't be visible again after restoring the wallet.
But the received transactions are returned correctly.

### Test plan

- Updated tests
- Manually tested sending with a fee currency adapter works:

| Image 1 | Image 2 | Image 3 |
|---------|---------|---------|
|
![image](https://github.com/valora-inc/wallet/assets/57791/e42b099c-4e18-45f6-be95-bf5816887717)
|
![image](https://github.com/valora-inc/wallet/assets/57791/a76f4b60-4457-43a7-9bd3-7583f349489b)
|
![image](https://github.com/valora-inc/wallet/assets/57791/e4604a06-8abf-4b92-9f8b-04957d6e19ea)
|
|
![image](https://github.com/valora-inc/wallet/assets/57791/13b7cbbb-35ef-44e6-9b5e-5944554afe23)
|
![image](https://github.com/valora-inc/wallet/assets/57791/be9f336d-0b5e-4962-a714-02426ebc8860)
| |

### Related issues

- Fixes RET-1009

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)