chore(deps): upgrade viem to ^2.7.16 (#5022)
### Description

Includes some typing changes.
- WalletActions need a more specific account type, it doesn't like the
`undefined` part of `Account | undefined`
- `baseFeePerGas` from `getBlock` returns `BigInt | null`, which I
believe has always been the case. In previous versions, we were getting
back `any` as the type which was assigned to `BigInt`. Now we throw if
it's null as it can only be null in blocks prior to the EIP-1559 upgrade

### Test plan

CI

### Related issues

- Fixes ACT-1111

### Backwards compatibility

Yes