fix: ensure screens that are bottom sheets scroll correctly (#5062)
### Description

Context
https://valora-app.slack.com/archives/C025V1D6F3J/p1709855852378389

The bottom sheet screens were not scrolling correctly, this was probably
a regression but unnoticed because most of these screens do not require
scrolling. Root cause is a known issue that is documented in a comment
in the Navigator - we need to use the react-native-gesture-handler
components to scroll inside bottom sheet screens.

This PR adds a prop to the BottomSheetScrollView component to use the
correct scrollview component.

### Test plan

Scrolling content:


https://github.com/valora-inc/wallet/assets/20150449/349c2628-017c-46b1-af88-06d20bcb19ce



Short content:


https://github.com/valora-inc/wallet/assets/20150449/5929f8f9-08f1-495f-8131-211cd055fbdb



Tested also on the WC flow.


### Related issues

- n/a

### Backwards compatibility

Y

### Network scalability

Y