feat(txs): Show Approval TXs only on supported chains (#5068)
### Description

For
[ACT-1081](https://linear.app/valora/issue/ACT-1081/use-showapprovaltxsinhomefeed-from-dynamic-config-in-wallet).
Adds an additional filter to the cached TXs selector to filter out
Approval TXs that aren't supported via Statsig.

### Test plan

Unit tested.

### Related issues

- Fixes
[ACT-1081](https://linear.app/valora/issue/ACT-1081/use-showapprovaltxsinhomefeed-from-dynamic-config-in-wallet).

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)