refactor(tokens): update networksIconSelector to return all icons (#5147)
### Description

Accepting network ids as an array would mean that the selector
recomputes the result even if the same list of network ids are passed
and returns a new result, resulting in unnecessary rerenders. This
updates it so that the selector returns all network id's icons. An
alternative approach would be to set `memoizeOptions` like
`tokenIdsSelector`, but there doesn't seem to be any harm in returning
all network's icon in the function. The consumer of this selector can
pick the required networks' icon from the object

### Test plan

CI, ensured the below warning doesn't show in #5148 when merged with
this change
<img
src="https://github.com/valora-inc/wallet/assets/5062591/14ac3745-d92c-44a7-9a1a-4c93837eff2e"
width="200"/>


### Related issues

N/A

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)