chore: remove name and picture and choose your adventure experiment (#5224)
### Description

Removes the Choose your own adventure (CYA) experiment code and makes
the screen permanent. Also removes the now unused name and picture
onboarding screen and FormInput component.


### Test plan

- Tested locally on iOS.
- E2E tests updated.
- Unit tests updated.

### Related issues

- Fixes act-1114

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)