chore(deps): remove react-native-fs custom fork resolution (#5320)
### Description

This is no longer needed. The custom fork only contained [2
commits](https://github.com/valora-inc/react-native-fs/commits/master/)
by our team, both changes are now included in the latest version of
`react-native-fs` (which we already have in the package.json).

### Test plan

n/a

### Related issues

- Related to RET-1058

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)