feat(points): Implement Points Activity bottom sheet cards (#5259)
### Description

Part of
[RET-1032](https://linear.app/valora/issue/RET-1032/[wallet-fake-backend]-points-activity-history).
Implements the points cards on the bottom sheet. Relies on the changes
in [this PR](https://github.com/valora-inc/points-functions/pull/44) in
the functions repo.

### Test plan

Unit and manual tested. See video.



https://github.com/valora-inc/wallet/assets/569401/675002c6-b74b-4c4b-8bc0-b5e809c4ee0c



### Related issues

- Fixes
[RET-1032](https://linear.app/valora/issue/RET-1032/[wallet-fake-backend]-points-activity-history)

### Backwards compatibility

Yes.

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)