ci: upload detox build to emerge on merge queue runs (#5360)
### Description

Some main runs fail to upload this because E2E-main runs are skipped if
multiple PRs merge at the same time (or if some fail due to test flake).
This ensures a commit on main always uploads a build to emerge.

More context here:
https://valora-app.slack.com/archives/C02E2FE98P2/p1713968251987719

### Test plan

Watch CI run on merge queue

### Related issues

N/A

### Backwards compatibility

N/A

### Network scalability

N/A