chore: simplify the NumberTicker interface (#5369)
### Description

As the title, this addresses some discussions we had in #5345 about
possible misuse of the props. This PR makes it harder to screw up the
styles.

### Test plan

No visual changes


https://github.com/valora-inc/wallet/assets/20150449/2b203175-13a2-4dc3-bbd5-e53923e82d5a


### Related issues

- Fixes RET-1072

### Backwards compatibility

Y
### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)