feat(earn): Add Crypto bottom sheet (#5376)
### Description

Creates new component for the Earn "Add Crypto" bottom sheet. This
component should be used in the same way as the TokenDetailsMoreActions
component [is
used](https://github.com/valora-inc/wallet/blob/6df3859e7271db5e643aba11eda8cde19151135d/src/tokens/TokenDetails.tsx#L119).

Decided to make a new component as to avoid adding extra props to decide
copy, choose actions, and add an amount prop which would not be used in
the TokenDetailsMoreActions case.

Talked with product and Nitya and confirmed that:
- On the QR code page, have title be "Deposit Crypto" like in the
cash-in exchange case
- Use "Arbitrum One" as the network name in copy

Note: Price is not passed to the swap component, there is not support
for that and based on discussion in quick sync it seems hard.

### Test plan

Unit tests added.

**Manual testing:**
Bottom sheet:

![bottom-sheet](https://github.com/valora-inc/wallet/assets/140328381/0a029234-6bba-46dd-8512-6a73a12b660a)

Tapping Buy:


https://github.com/valora-inc/wallet/assets/140328381/2a8985ca-9ae2-41e1-b0f7-abed4071b723

Tapping Transfer:


https://github.com/valora-inc/wallet/assets/140328381/61e38183-d641-43c8-9d2d-39af4236215d



Tapping Swap:


https://github.com/valora-inc/wallet/assets/140328381/393d00a6-41d7-49fc-a0f7-162e1ed15e71


Tapping 

### Related issues

- Fixes #ACT-1177

### Backwards compatibility

Yes, new component so not changing past stuff

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [X] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)