feat: add earn active pool component (#5400)
### Description

Creates the `EarnActivePool.tsx` component and associated `useEarn`
hook. This hook is then used on the `TabDiscover.tsx` screen to display
either `EarnActivePool.tsx` or `EarnCta.tsx`.

#### Video iOS


https://github.com/valora-inc/wallet/assets/26950305/70528be7-f239-4c7b-9e65-a97409520647

### Test plan

- [x] Tested locally on iOS
- [x] Tested locally on Android
- [x] Unit tests added

### Related issues

- ACT-1179

### Backwards compatibility

Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)